requirements.txt
google-genai>=0.3.0
pytz>=2024.1
